remove text size
