+++ packages/radix-ui-themes/src/components/text.props.ts
@@ -1,15 +1,23 @@
 import { weightProp, alignProp, trimProp, colorProp, highContrastProp } from '../helpers';
-import { textSize } from '../helpers/props/text-size.prop';
+import type { PropDef } from '../helpers';
+
+const sizes = ['1', '2', '3', '4', '5', '6', '7', '8', '9'] as const;
 
 const textPropDefs = {
-  size: textSize,
+  size: {
+    type: 'enum',
+    className: 'rt-r-size',
+    values: sizes,
+    default: undefined,
+    responsive: true,
+  },
   weight: weightProp,
   align: alignProp,
   trim: trimProp,
   color: colorProp,
   highContrast: highContrastProp,
 } satisfies {
-  size: typeof textSize;
+  size: PropDef<(typeof sizes)[number]>;
   weight: typeof weightProp;
   align: typeof alignProp;
   trim: typeof trimProp;
