--- packages/radix-ui-themes/src/helpers/props/text-size.prop.ts
@@ -1,13 +1,0 @@
-import type { PropDef } from '..';
-
-const sizes = ['1', '2', '3', '4', '5', '6', '7', '8', '9'] as const;
-
-const textSize = {
-  type: 'enum',
-  className: 'rt-r-size',
-  values: sizes,
-  default: undefined,
-  responsive: true,
-} satisfies PropDef<(typeof sizes)[number]>;
-
-export { textSize };
